% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firasans-package.R
\docType{package}
\name{firasans}
\alias{firasans}
\alias{firasans-package}
\title{Fira Sans Condensed / Fira Mono Font Theme Based on hrbrthemes}
\description{
Fira Sans Condensed / Fira Mono Font Theme Based on hrbrthemes
}
\author{
Bob Rudis (bob@rud.is)
}
