% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firasans.R
\docType{data}
\name{font_fsc}
\alias{font_fsc}
\alias{font_fsm}
\title{Fira Sans Condensed font name R variable aliases}
\format{length 1 character vector}
\usage{
font_fsc

font_fsm
}
\description{
\code{font_fsc} == "\code{Fira Sans Condensed}"

\code{font_fsm} == "\code{Fira Mono}"
}
\keyword{datasets}
